backend/requirements.txt
# Core API
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.6.0
pydantic-settings>=2.1.0
python-dotenv>=1.0.0

# AI & LangChain
langchain>=0.1.0
langchain-community>=0.0.16
langchain-core>=0.1.20
langchain-groq>=0.0.1
sentence-transformers>=2.3.0

# Vector Database
chromadb>=0.4.22
langchain-chroma

# Data Processing
pandas>=2.2.0
numpy>=1.26.0